Adding APIS for Players and refactoring Hiberbate many to many relationship

+++ lab2/src/main/java/com/cmpe275/lab2/service/PlayerService.java
@@ -9,6 +9,6 @@ public interface PlayerService {
 	
 	public Player createPlayer(Player player,String sponsorName);
 	public Player updatePlayer(Player player,String sponsorName);
-	public Player getPlayer(int id);
-	public Player deletePlayer(int id);
+	public Player getPlayer(long id);
+	public Player deletePlayer(long id);
 }
